pyproject.toml
[tool.poetry]
name = "udali"
version = "0.1.0"
description = ""
authors = ["Kenesary <kenesarysapargali@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-storage = "^3.0.0"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
